Rearrange api helper functions to get rid of a catch-all utils file
